Replaced "int" Ids by "long" Ids.

+++ src/fr/epf/computer/dao/CompanyDao.java
@@ -10,6 +10,6 @@ public interface CompanyDao {
     List<Company> getCompanies();
     List<Company> searchByName(String name);
     void create(Company company);
-    Company findById(int id);
+    Company findById(long id);
 
 }
