+++ src/fr/epf/computer/domain/Company.java
@@ -12,11 +12,11 @@ public class Company {
 
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -30,7 +30,7 @@ public class Company {
 
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
 
     @Column(name = "name")
     private String name;
@@ -42,7 +42,7 @@ public class Company {
             company = new Company();
         }
 
-        public Builder id(int id) {
+        public Builder id(long id) {
             company.setId(id);
             return this;
         }
