+++ src/fr/epf/computer/domain/Computer.java
@@ -13,7 +13,7 @@ public class Computer {
 
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
 
     @Column( name = "name")
     private String name;
@@ -35,11 +35,11 @@ public class Computer {
 
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -82,7 +82,7 @@ public class Computer {
 			computer = new Computer();
 		}
 		
-		public Builder id(int id) {
+		public Builder id(long id) {
 			computer.setId(id);
 			return this;
 		}
